+++ src/main/java/team5/proyecto/reservesMenjador/dao/IRolDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import team5.proyecto.reservesMenjador.dto.Rol;
 
-public interface IRolDAO extends JpaRepository<Rol, Long>{
+public interface IRolDAO extends JpaRepository<Rol, Integer>{
 
 }
